+++ packages/astro/src/actions/runtime/utils.ts
@@ -1,3 +1,6 @@
+import type { ZodType } from 'zod';
+import type { ActionAccept, ActionClient } from './virtual/server.js';
+
 export const formContentTypes = ['application/x-www-form-urlencoded', 'multipart/form-data'];
 
 export function hasContentType(contentType: string, expected: string[]) {
@@ -12,14 +15,16 @@ export type MaybePromise<T> = T | Promise<T>;
 
 /**
  * Get server-side action based on the route path.
- * Imports from `import.meta.env.ACTIONS_PATH`, which maps to
+ * Imports from the virtual module `astro:internal-actions`, which maps to
  * the user's `src/actions/index.ts` file at build-time.
  */
 export async function getAction(
 	path: string
-): Promise<((param: unknown) => MaybePromise<unknown>) | undefined> {
+): Promise<ActionClient<unknown, ActionAccept, ZodType> | undefined> {
 	const pathKeys = path.replace('/_actions/', '').split('.');
-	let { server: actionLookup } = await import(import.meta.env.ACTIONS_PATH);
+	// @ts-expect-error virtual module
+	let { server: actionLookup } = await import('astro:internal-actions');
+
 	for (const key of pathKeys) {
 		if (!(key in actionLookup)) {
 			return undefined;
@@ -31,3 +36,14 @@ export async function getAction(
 	}
 	return actionLookup;
 }
+
+/**
+ * Used to preserve the input schema type in the error object.
+ * This allows for type inference on the `fields` property
+ * when type narrowed to an `ActionInputError`.
+ *
+ * Example: Action has an input schema of `{ name: z.string() }`.
+ * When calling the action and checking `isInputError(result.error)`,
+ * `result.error.fields` will be typed with the `name` field.
+ */
+export type ErrorInferenceObject = Record<string, any>;