+++ packages/astro/src/actions/runtime/virtual/server.ts
@@ -1,13 +1,7 @@
 import { z } from 'zod';
 import { type ActionAPIContext, getApiContext as _getApiContext } from '../store.js';
-import { type MaybePromise } from '../utils.js';
-import {
-	ActionError,
-	ActionInputError,
-	type ErrorInferenceObject,
-	type SafeResult,
-	callSafely,
-} from './shared.js';
+import type { ErrorInferenceObject, MaybePromise } from '../utils.js';
+import { ActionError, ActionInputError, type SafeResult, callSafely } from './shared.js';
 
 export * from './shared.js';
 
@@ -16,42 +10,45 @@ export { z } from 'zod';
 /** @deprecated Access context from the second `handler()` parameter. */
 export const getApiContext = _getApiContext;
 
-export type Accept = 'form' | 'json';
-export type InputSchema<T extends Accept> = T extends 'form'
+export type ActionAccept = 'form' | 'json';
+export type ActionInputSchema<T extends ActionAccept | undefined> = T extends 'form'
 	? z.AnyZodObject | z.ZodType<FormData>
 	: z.ZodType;
 
-type Handler<TInputSchema, TOutput> = TInputSchema extends z.ZodType
+export type ActionHandler<TInputSchema, TOutput> = TInputSchema extends z.ZodType
 	? (input: z.infer<TInputSchema>, context: ActionAPIContext) => MaybePromise<TOutput>
 	: (input: any, context: ActionAPIContext) => MaybePromise<TOutput>;
 
+export type ActionReturnType<T extends ActionHandler<any, any>> = Awaited<ReturnType<T>>;
+
 export type ActionClient<
 	TOutput,
-	TAccept extends Accept,
-	TInputSchema extends InputSchema<TAccept> | undefined,
+	TAccept extends ActionAccept | undefined,
+	TInputSchema extends ActionInputSchema<TAccept> | undefined,
 > = TInputSchema extends z.ZodType
 	? ((
 			input: TAccept extends 'form' ? FormData : z.input<TInputSchema>
-		) => Promise<Awaited<TOutput>>) & {
-			safe: (
+		) => Promise<
+			SafeResult<
+				z.input<TInputSchema> extends ErrorInferenceObject
+					? z.input<TInputSchema>
+					: ErrorInferenceObject,
+				Awaited<TOutput>
+			>
+		>) & {
+			queryString: string;
+			orThrow: (
 				input: TAccept extends 'form' ? FormData : z.input<TInputSchema>
-			) => Promise<
-				SafeResult<
-					z.input<TInputSchema> extends ErrorInferenceObject
-						? z.input<TInputSchema>
-						: ErrorInferenceObject,
-					Awaited<TOutput>
-				>
-			>;
+			) => Promise<Awaited<TOutput>>;
 		}
-	: ((input?: any) => Promise<Awaited<TOutput>>) & {
-			safe: (input?: any) => Promise<SafeResult<never, Awaited<TOutput>>>;
+	: (input?: any) => Promise<SafeResult<never, Awaited<TOutput>>> & {
+			orThrow: (input?: any) => Promise<Awaited<TOutput>>;
 		};
 
 export function defineAction<
 	TOutput,
-	TAccept extends Accept = 'json',
-	TInputSchema extends InputSchema<Accept> | undefined = TAccept extends 'form'
+	TAccept extends ActionAccept | undefined = undefined,
+	TInputSchema extends ActionInputSchema<ActionAccept> | undefined = TAccept extends 'form'
 		? // If `input` is omitted, default to `FormData` for forms and `any` for JSON.
 			z.ZodType<FormData>
 		: undefined,
@@ -62,23 +59,26 @@ export function defineAction<
 }: {
 	input?: TInputSchema;
 	accept?: TAccept;
-	handler: Handler<TInputSchema, TOutput>;
-}): ActionClient<TOutput, TAccept, TInputSchema> {
+	handler: ActionHandler<TInputSchema, TOutput>;
+}): ActionClient<TOutput, TAccept, TInputSchema> & string {
 	const serverHandler =
 		accept === 'form'
 			? getFormServerHandler(handler, inputSchema)
 			: getJsonServerHandler(handler, inputSchema);
 
-	Object.assign(serverHandler, {
-		safe: async (unparsedInput: unknown) => {
-			return callSafely(() => serverHandler(unparsedInput));
-		},
+	const safeServerHandler = async (unparsedInput: unknown) => {
+		return callSafely(() => serverHandler(unparsedInput));
+	};
+
+	Object.assign(safeServerHandler, {
+		orThrow: serverHandler,
 	});
-	return serverHandler as ActionClient<TOutput, TAccept, TInputSchema>;
+
+	return safeServerHandler as ActionClient<TOutput, TAccept, TInputSchema> & string;
 }
 
-function getFormServerHandler<TOutput, TInputSchema extends InputSchema<'form'>>(
-	handler: Handler<TInputSchema, TOutput>,
+function getFormServerHandler<TOutput, TInputSchema extends ActionInputSchema<'form'>>(
+	handler: ActionHandler<TInputSchema, TOutput>,
 	inputSchema?: TInputSchema
 ) {
 	return async (unparsedInput: unknown): Promise<Awaited<TOutput>> => {
@@ -99,8 +99,8 @@ function getFormServerHandler<TOutput, TInputSchema extends InputSchema<'form'>>
 	};
 }
 
-function getJsonServerHandler<TOutput, TInputSchema extends InputSchema<'json'>>(
-	handler: Handler<TInputSchema, TOutput>,
+function getJsonServerHandler<TOutput, TInputSchema extends ActionInputSchema<'json'>>(
+	handler: ActionHandler<TInputSchema, TOutput>,
 	inputSchema?: TInputSchema
 ) {
 	return async (unparsedInput: unknown): Promise<Awaited<TOutput>> => {
@@ -128,8 +128,8 @@ export function formDataToObject<T extends z.AnyZodObject>(
 	const obj: Record<string, unknown> = {};
 	for (const [key, baseValidator] of Object.entries(schema.shape)) {
 		let validator = baseValidator;
-		if (baseValidator instanceof z.ZodOptional || baseValidator instanceof z.ZodNullable) {
-			validator = baseValidator._def.innerType;
+		while (validator instanceof z.ZodOptional || validator instanceof z.ZodNullable) {
+			validator = validator._def.innerType;
 		}
 		if (validator instanceof z.ZodBoolean) {
 			obj[key] = formData.has(key);