+++ packages/astro/src/core/render-context.ts
@@ -21,6 +21,8 @@ import { renderPage } from '../runtime/server/index.js';
 import {
 	ASTRO_VERSION,
 	REROUTE_DIRECTIVE_HEADER,
+	REWRITE_DIRECTIVE_HEADER_KEY,
+	REWRITE_DIRECTIVE_HEADER_VALUE,
 	ROUTE_TYPE_HEADER,
 	clientAddressSymbol,
 	clientLocalsSymbol,
@@ -146,6 +148,7 @@ export class RenderContext {
 					this.routeData = routeData;
 					componentInstance = component;
 					this.isRewriting = true;
+					this.status = 200;
 				} else {
 					this.pipeline.logger.error(
 						'router',
@@ -183,12 +186,11 @@ export class RenderContext {
 					// Signal to the i18n middleware to maybe act on this response
 					response.headers.set(ROUTE_TYPE_HEADER, 'page');
 					// Signal to the error-page-rerouting infra to let this response pass through to avoid loops
-					if (
-						this.routeData.route === '/404' ||
-						this.routeData.route === '/500' ||
-						this.isRewriting
-					) {
+					if (this.routeData.route === '/404' || this.routeData.route === '/500') {
 						response.headers.set(REROUTE_DIRECTIVE_HEADER, 'no');
+					}
+					if (this.isRewriting) {
+						response.headers.set(REWRITE_DIRECTIVE_HEADER_KEY, REWRITE_DIRECTIVE_HEADER_VALUE);
 					}
 					break;
 				}
@@ -230,6 +232,43 @@ export class RenderContext {
 		});
 	}
 
+	async #executeRewrite(reroutePayload: RewritePayload) {
+		this.pipeline.logger.debug('router', 'Calling rewrite: ', reroutePayload);
+		if (!this.pipeline.manifest.rewritingEnabled) {
+			this.pipeline.logger.error(
+				'router',
+				'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.'
+			);
+			return new Response(
+				'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
+				{
+					status: 500,
+					statusText:
+						'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
+				}
+			);
+		}
+		const [routeData, component, newURL] = await this.pipeline.tryRewrite(
+			reroutePayload,
+			this.request,
+			this.originalRoute
+		);
+		this.routeData = routeData;
+		if (reroutePayload instanceof Request) {
+			this.request = reroutePayload;
+		} else {
+			this.request = this.#copyRequest(newURL, this.request);
+		}
+		this.url = new URL(this.request.url);
+		this.cookies = new AstroCookies(this.request);
+		this.params = getParams(routeData, this.url.pathname);
+		this.pathname = this.url.pathname;
+		this.isRewriting = true;
+		// we found a route and a component, we can change the status code to 200
+		this.status = 200;
+		return await this.render(component);
+	}
+
 	createActionAPIContext(): ActionAPIContext {
 		const renderContext = this;
 		const { cookies, params, pipeline, url } = this;
@@ -238,38 +277,7 @@ export class RenderContext {
 			new Response(null, { status, headers: { Location: path } });
 
 		const rewrite = async (reroutePayload: RewritePayload) => {
-			pipeline.logger.debug('router', 'Called rewriting to:', reroutePayload);
-			if (!this.pipeline.manifest.rewritingEnabled) {
-				this.pipeline.logger.error(
-					'router',
-					'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.'
-				);
-				return new Response(
-					'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
-					{
-						status: 500,
-						statusText:
-							'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
-					}
-				);
-			}
-			const [routeData, component, newURL] = await pipeline.tryRewrite(
-				reroutePayload,
-				this.request,
-				this.originalRoute
-			);
-			this.routeData = routeData;
-			if (reroutePayload instanceof Request) {
-				this.request = reroutePayload;
-			} else {
-				this.request = this.#copyRequest(newURL, this.request);
-			}
-			this.url = newURL;
-			this.cookies = new AstroCookies(this.request);
-			this.params = getParams(routeData, this.url.pathname);
-			this.isRewriting = true;
-			this.pathname = this.url.pathname;
-			return await this.render(component);
+			return await this.#executeRewrite(reroutePayload);
 		};
 
 		return {
@@ -339,6 +347,7 @@ export class RenderContext {
 		// This object starts here as an empty shell (not yet the result) but then
 		// calling the render() function will populate the object with scripts, styles, etc.
 		const result: SSRResult = {
+			base: manifest.base,
 			cancelled: false,
 			clientDirectives,
 			inlinedScripts,
@@ -349,6 +358,7 @@ export class RenderContext {
 			createAstro: (astroGlobal, props, slots) =>
 				this.createAstro(result, astroGlobal, props, slots),
 			links,
+			params: this.params,
 			partial,
 			pathname,
 			renderers,
@@ -358,6 +368,8 @@ export class RenderContext {
 			scripts,
 			styles,
 			actionResult,
+			serverIslandNameMap: manifest.serverIslandNameMap ?? new Map(),
+			trailingSlash: manifest.trailingSlash,
 			_metadata: {
 				hasHydrationScript: false,
 				rendererSpecificHydrationScripts: new Set(),
@@ -374,6 +386,7 @@ export class RenderContext {
 	}
 
 	#astroPagePartial?: Omit<AstroGlobal, 'props' | 'self' | 'slots'>;
+
 	/**
 	 * The Astro global is sourced in 3 different phases:
 	 * - **Static**: `.generator` and `.glob` is printed by the compiler, instantiated once per process per astro file
@@ -447,38 +460,7 @@ export class RenderContext {
 		};
 
 		const rewrite = async (reroutePayload: RewritePayload) => {
-			if (!this.pipeline.manifest.rewritingEnabled) {
-				this.pipeline.logger.error(
-					'router',
-					'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.'
-				);
-				return new Response(
-					'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
-					{
-						status: 500,
-						statusText:
-							'The rewrite API is experimental. To use this feature, add the `rewriting` flag to the `experimental` object in your Astro config.',
-					}
-				);
-			}
-			pipeline.logger.debug('router', 'Calling rewrite: ', reroutePayload);
-			const [routeData, component, newURL] = await pipeline.tryRewrite(
-				reroutePayload,
-				this.request,
-				this.originalRoute
-			);
-			this.routeData = routeData;
-			if (reroutePayload instanceof Request) {
-				this.request = reroutePayload;
-			} else {
-				this.request = this.#copyRequest(newURL, this.request);
-			}
-			this.url = new URL(this.request.url);
-			this.cookies = new AstroCookies(this.request);
-			this.params = getParams(routeData, this.url.pathname);
-			this.pathname = this.url.pathname;
-			this.isRewriting = true;
-			return await this.render(component);
+			return await this.#executeRewrite(reroutePayload);
 		};
 
 		return {
@@ -536,6 +518,7 @@ export class RenderContext {
 	 * So, it is computed and saved here on creation of the first APIContext and reused for later ones.
 	 */
 	#currentLocale: APIContext['currentLocale'];
+
 	computeCurrentLocale() {
 		const {
 			url,
@@ -561,6 +544,7 @@ export class RenderContext {
 	}
 
 	#preferredLocale: APIContext['preferredLocale'];
+
 	computePreferredLocale() {
 		const {
 			pipeline: { i18n },
@@ -571,6 +555,7 @@ export class RenderContext {
 	}
 
 	#preferredLocaleList: APIContext['preferredLocaleList'];
+
 	computePreferredLocaleList() {
 		const {
 			pipeline: { i18n },