+++ packages/astro/src/core/routing/astro-designed-error-pages.ts
@@ -1,4 +1,4 @@
-import type { ManifestData, RouteData } from '../../@types/astro.js';
+import type { ComponentInstance, ManifestData, RouteData } from '../../@types/astro.js';
 import notFoundTemplate from '../../template/4xx.js';
 import { DEFAULT_404_COMPONENT, DEFAULT_500_COMPONENT } from '../constants.js';
 
@@ -37,7 +37,7 @@ export function ensure404Route(manifest: ManifestData) {
 	return manifest;
 }
 
-export async function default404Page({ pathname }: { pathname: string }) {
+async function default404Page({ pathname }: { pathname: string }) {
 	return new Response(
 		notFoundTemplate({
 			statusCode: 404,
@@ -50,3 +50,7 @@ export async function default404Page({ pathname }: { pathname: string }) {
 }
 // mark the function as an AstroComponentFactory for the rendering internals
 default404Page.isAstroComponentFactory = true;
+
+export const default404Instance: ComponentInstance = {
+	default: default404Page,
+};